remove spacesuit.rs: prove/verify are only for tests
