--- src/gadgets/mod.rs
@@ -1,12 +1,0 @@
-pub mod transaction;
-
-mod k_mix;
-mod merge;
-mod mix;
-mod split;
-
-mod padded_shuffle;
-mod scalar_shuffle;
-mod value_shuffle;
-
-mod range_proof;
