+++ src/lib.rs
@@ -4,9 +4,24 @@ extern crate merlin;
 extern crate rand;
 extern crate subtle;
 
-mod gadgets;
-mod spacesuit;
+mod k_mix;
+mod merge;
+mod mix;
+mod split;
+
+mod padded_shuffle;
+mod scalar_shuffle;
+mod value_shuffle;
+
+mod range_proof;
+mod cloak;
+
 mod value;
 
-pub use spacesuit::*;
-pub use value::*;
+pub use cloak::cloak;
+pub use range_proof::range_proof;
+pub use value::{Value, CommittedValue};
+
+// TBD: figure out if we need to export these at all
+pub use value::{ProverCommittable, VerifierCommittable};
+
