+++ tests/spacesuit.rs
@@ -1,8 +1,14 @@
-extern crate spacesuit;
-use spacesuit::{prove, verify, Value};
+extern crate rand;
+extern crate merlin;
 extern crate bulletproofs;
-use bulletproofs::r1cs::R1CSError;
+extern crate spacesuit;
+
+use rand::{CryptoRng, Rng};
+use bulletproofs::r1cs::{Prover, R1CSError, R1CSProof, Verifier};
 use bulletproofs::{BulletproofGens, PedersenGens};
+use merlin::Transcript;
+
+use spacesuit::{cloak, CommittedValue, ProverCommittable, Value, VerifierCommittable};
 
 fn spacesuit_helper(
     bp_gens: &BulletproofGens,
@@ -15,6 +21,47 @@ fn spacesuit_helper(
     let (proof, in_com, out_com) = prove(&bp_gens, &pc_gens, &inputs, &outputs, &mut rng)?;
 
     verify(&bp_gens, &pc_gens, &proof, &in_com, &out_com)
+}
+
+fn prove<R: Rng + CryptoRng>(
+    bp_gens: &BulletproofGens,
+    pc_gens: &PedersenGens,
+    inputs: &Vec<Value>,
+    outputs: &Vec<Value>,
+    rng: &mut R,
+) -> Result<(R1CSProof, Vec<CommittedValue>, Vec<CommittedValue>), R1CSError>
+where
+    R: rand::RngCore,
+{
+    let mut prover_transcript = Transcript::new(b"TransactionTest");
+    let mut prover = Prover::new(&bp_gens, &pc_gens, &mut prover_transcript);
+
+    let (in_com, in_vars) = inputs.commit(&mut prover, rng);
+    let (out_com, out_vars) = outputs.commit(&mut prover, rng);
+
+    cloak(&mut prover, in_vars, out_vars)?;
+    let proof = prover.prove()?;
+
+    Ok((proof, in_com, out_com))
+}
+
+fn verify(
+    bp_gens: &BulletproofGens,
+    pc_gens: &PedersenGens,
+    proof: &R1CSProof,
+    in_com: &Vec<CommittedValue>,
+    out_com: &Vec<CommittedValue>,
+) -> Result<(), R1CSError> {
+    // Verifier makes a `ConstraintSystem` instance representing a merge gadget
+    let mut verifier_transcript = Transcript::new(b"TransactionTest");
+    let mut verifier = Verifier::new(&bp_gens, &pc_gens, &mut verifier_transcript);
+
+    let in_vars = in_com.commit(&mut verifier);
+    let out_vars = out_com.commit(&mut verifier);
+
+    assert!(cloak(&mut verifier, in_vars, out_vars,).is_ok());
+
+    Ok(verifier.verify(&proof)?)
 }
 
 // Helper functions to make the tests easier to read
